+++ pyfragment/__init__.py
@@ -7,8 +7,7 @@ from pyfragment.domains.anonymous_numbers.models import LoginCodeResult, Termina
 from pyfragment.domains.giveaways.models import PremiumGiveawayResult, StarsGiveawayResult
 from pyfragment.domains.marketplace.models import GiftsResult, NumbersResult, UsernamesResult
 from pyfragment.domains.purchases.models import PremiumResult, StarsResult
-from pyfragment.domains.tonapi.models import WalletInfo
-from pyfragment.enums import PaymentMethod, WalletVersion
+from pyfragment.enums import ApiProvider, PaymentMethod, WalletVersion
 from pyfragment.exceptions import (
     AlreadySubscribedError,
     AnonymousNumberError,
@@ -27,6 +26,7 @@ from pyfragment.exceptions import (
     WalletError,
 )
 from pyfragment.services.cookies import CookieResult, get_cookies_from_browser
+from pyfragment.services.tonapi.models import WalletInfo
 
 logging.getLogger("pyfragment").addHandler(logging.NullHandler())
 
@@ -66,6 +66,7 @@ __all__ = [
     "ParseError",
     "UnexpectedError",
     # literal types
+    "ApiProvider",
     "PaymentMethod",
     "WalletVersion",
     "get_cookies_from_browser",