+++ pyfragment/domains/giveaways/giveaway.py
@@ -19,8 +19,6 @@ from pyfragment.core.constants import (
 )
 from pyfragment.domains.giveaways.models import PremiumGiveawayResult, StarsGiveawayResult
 from pyfragment.domains.payments import parse_required_payment_amount
-from pyfragment.domains.tonapi.account import get_account_info
-from pyfragment.domains.tonapi.transaction import process_transaction
 from pyfragment.enums import PaymentMethod
 from pyfragment.exceptions import (
     ConfigurationError,
@@ -30,6 +28,8 @@ from pyfragment.exceptions import (
     UserNotFoundError,
     VerificationError,
 )
+from pyfragment.services.tonapi.account import get_account_info
+from pyfragment.services.tonapi.transaction import process_transaction
 
 if TYPE_CHECKING:
     from pyfragment.client import FragmentClient