+++ pyfragment/domains/tonapi/__init__.py
@@ -1,5 +1,0 @@
-from pyfragment.domains.tonapi.service import TonapiService
-
-__all__ = [
-    "TonapiService",
-]